Use MaxLiquiditySpread enum in UI
